# Small optimization to PrometheusResponse marshalling